Forgot to commit last night
- Changed the edge radius of the obstacles from 0 to 0.1
- Changed the mass from 1 to 0.1
- Changed some sorting orders of a few objects within the scene I was testing in
- Added some more points for the obstacles to between for the purpose of testing more
