+++ 00 Unity Proj/Bubble/Assets/Scripts/Game/PlayerRespawn.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 67f1f565610265043a6dbd1b64990ed1
+guid: b25a631fdb87db046a77c6e066db5bee
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
